+++ pkgs/http/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.1
+* `BrowserClient` throws `ClientException` when the `'Content-Length'` header
+  is invalid.
+* `IOClient` trims trailing whitespace on header values.
+
 ## 1.1.0
 
 * Add better error messages for `SocketException`s when using `IOClient`.